move SpeechRecognizerProtocol.swift to its own file
